requirements.txt
flask==3.0.0
flask-cors==4.0.0
facenet-pytorch==2.5.3
torch==2.1.2
torchvision==0.16.2
numpy==1.26.4
pillow==10.1.0
opencv-python-headless==4.9.0.80
gunicorn==21.2.0
requests==2.31.0
